refactor: use CollapsedPopup
